Mostly BGBCC tweaks...
Working on bit handling and Verilog support in BGBCC.
Also took a detour into resource handling stuff, but did end up tweaking the layout of the default color palette...

+++ bgbcc22/mm/cca_node.c
@@ -654,7 +654,7 @@ char *BCCX_GetCst(BCCX_Node *n,
 		if((*an>>12)==BCCX_IVTY_INT)
 		{
 			tb=bgbcc_ralloc(64);
-			sprintf(tb, "%lld", av->i);
+			sprintf(tb, "%lld", (long long)(av->i));
 			return(tb);
 		}
